scripts/pyproject.toml
[tool.poetry]
name = "scripts"
version = "0.1.0"
description = "Scripts for use with the thesis tracker"
authors = ["George Kaye <georgejkaye@gmail.com>"]
license = "GPL-v3.0-or-later"
readme = "README.md"
packages = [{include = "scripts", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
gitpython = "^3.1.40"

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
mypy = "^1.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
